requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asteroid-filterbanks==0.4.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
deep-translator==1.11.4
demucs==4.0.1
distlib==0.4.0
docopt==0.6.2
dora_search==0.1.12
einops==0.8.1
fastapi==0.116.1
ffmpeg==1.4
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
googletrans==4.0.0rc1
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
hf-xet==1.1.5
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.34.1
hyperframe==5.2.0
HyperPyYAML==1.2.2
idna==2.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
joblib==1.5.1
julius==0.2.7
kiwisolver==1.4.8
lameenc==1.8.1
lightning==2.5.2
lightning-utilities==0.14.3
llvmlite==0.44.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
moviepy==2.2.1
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
numba==0.61.2
numpy==2.2.6
omegaconf==2.3.0
openai-whisper==20250625
openunmix==1.3.0
optuna==4.4.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
primePy==1.3
proglog==0.1.12
propcache==0.3.2
protobuf==6.31.1
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-lightning==2.5.2
pytorch-metric-learning==2.8.1
pytube==15.0.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
retrying==1.4.1
rfc3986==1.5.0
rich==14.1.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-learn==1.7.1
scipy==1.16.0
semver==3.0.4
sentencepiece==0.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
speechbrain==1.0.3
SQLAlchemy==2.0.41
srt==3.5.3
starlette==0.47.2
submitit==1.5.3
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.7.1
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.1
torchmetrics==1.8.0
torchvision==0.22.1
tqdm==4.67.1
treetable==0.2.5
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.32.0
yarl==1.20.1
yt-dlp==2025.7.21
